user controller updated

+++ src/main/java/az/edu/turing/booking/service/impl/UserServiceImpl.java
@@ -11,6 +11,7 @@ import az.edu.turing.booking.model.dto.request.UserRoleUpdateRequest;
 import az.edu.turing.booking.model.dto.request.UsernameUpdateRequest;
 import az.edu.turing.booking.model.enums.UserRole;
 import az.edu.turing.booking.service.UserService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -80,10 +81,15 @@ public class UserServiceImpl implements UserService {
                 .orElseThrow(() -> new NotFoundException("User not found with username: " + username));
     }
 
+    @Transactional
     @Override
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User with id " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
+
 
     @Override
     public boolean isAdmin(Long id) {
